Added new scene
Added a new scene - Main Menu.
Changed the structure of the code so that one script will hold the score, like a manager script.
Added to that new scene some new UI like buttons and text.
You can go from the game itself and to Main Menu and start the game again.